+++ history.md
@@ -4,6 +4,10 @@
 
 ### AzGovViz version 6
 
+__Changes__ (2021-Oct-25 / Major)
+
+* AzAPICall enhanced error handling (general error 'An error has occurred.' ; roleAssignment schedules)
+
 __Changes__ (2021-Oct-21 / Major)
 
 * AzAPICall enhanced error handling (GatewayAuthenticationFailed; roleAssignment schedules)